Extract shared file context behaviour
